Refactor Logger.get()
This allows Logger.get() to be queried for a specific number of log entries. Additionally, it improves the control over the method, replacing the String argmuent with a boolean one that determines whether entries should start at the beginning or end of the log. This removes the need for a warning and significantly simplifies the code overall.