Add Categories.PRINT and Categories.SAVE
Categories.PRINT replaces the now-deprecated Categories.ENABLED, printing a log entry but not saving it. Categories.SAVE indicates an entry should be saved to the log file but not printed. Categories.ENABLED can still be used but will be removed in later versions. This commit also updates an out-of-date docstring for the Logger class.